BitMart 1

+++ CoinsLoader/Model/Chain.cs
@@ -6,11 +6,11 @@ public class Chain
     public int coinId { get; set; }
     public string? chainName { get; set; }
     public string? contractAddress { get; set; }
-    public float? withdrawFee { get; set; }
+    public double? withdrawFee { get; set; }
     public bool? allowDeposit { get; set; }
     public bool? allowWithdraw { get; set; }
-    public float? minDepositAmt { get; set; }
-    public float? minWithdrawal { get; set; }
+    public double? minDepositAmt { get; set; }
+    public double? minWithdrawal { get; set; }
     public DateTime? dtu { get; set; }
     public async Task Save()
     {
